NetworkConfig

+++ NetworkConfig.py
@@ -1,7 +1,7 @@
 import numpy as np
 
 class NetworkConfig:
-    DEFAULT_CAPACITY_NETWORK = np.asarray([4096,4096,4096,4096,4096])
+    DEFAULT_CAPACITY_NETWORK = np.asarray([4096,4096,4096,4096,4096], dtype=np.int64)
